Proposed updates for readability and less code
I combined the two macros into one and added support for a `type` attribute to be used. The macro supports xhtml and text blocks with this version. I found an issue with how text blocks are handled in the Cascade API. I already made the code change and submitted it for review. Hopefully it will go out in the next patch release. Until then it's necessary to strip the `<system-xml>` tags from the output of a Text Block. I also added various commas and spaces for readability. And lastly I included more code for accessing the WYSIWYG on the page via the Data Definition so it was more clear where the content was coming from. Let me know what you think.